Save and restore point in copy mode.
This change will save the point when entering copy mode and restore it back once
leaving. This feels much nicer than having the point get out of sync with where
the terminal actually thinks it is.

If the creators of vterm know of a better way I would be glad to figure out how
to implement it, but for now this seems to work fine.