This commit fixes the newlines when copied and pasted.
This commit fixes up the newlines that were inserted for wrapping using the same
approach as term.el does. It does so by adding a property to the '\n' chars that
were inserted as a result of wrapping the terminal. When killing a region text
will be passed through the filter-buffer-substring-function
vterm--remove-fake-newlines removing the wrapped new lines by looking for the
'vterm-line-wrap property and removing the char that corresponds.

If this change set is not up to project standards please work with me to correct
any issues you find.