vterm-copy-mode-done copies line when no region is set
When a region is not selected the current line will be killed.

The option `vterm-copy-exclude-prompt` is used to determine if the prompt should be excluded from the copy.

When `vterm-copy-use-vterm-prompt` is true it will use the built in vterm prompt tracking functions to skip the prompt.

If prompt tracking doesn't work you can set it to `nil` and use `vterm-copy-prompt-regexp` to search for the end of the
prompt.