[raxod502/straight.el#504] Don't resolve symlinks
Under package.el, the package files are not symlinks. However, under
straight.el, the package files are originally in the Git repository
but they are symlinked into a build directory, which is then added to
the load-path. Calling file-truename on the vterm library during the
native module compilation causes the native module to be put into the
Git repository rather than the build directory, meaning it cannot be
found by Emacs.