First build of continuous integration for emacs-libvterm
This builds on `ubuntu-latest` and `macos-latest`.

Using github actions this:
- retrieves the source
- installs any dependencies (libtool for ubuntu)
- Runs the install as described in README.md
  - makes the Build directory
  - enters it and runs `cmake ..`, returns to home
  - enters it and runs `make`, returns to home

When tests are added we can added `make test` step as well.

I left it as only running on the vterm-ci branch until it's decided how this will fit into the workflow (e.g. does it run on a testing branch or master or both?)

Building on `windows-latest` looks like it should be possible, the image includes cmake support, but I haven't used windows in ages and couldn't figure out how to install libtool.  Someone with more recent experience might have better luck.