Fix a recent gross hack
Fix [1: beda920], whose commit message was quite accurate.  That
commit used `defvar-local' in the believe that this would define a
variable with some sort of local scope, which would be the correct
thing to do, but unfortunately that's not what this macro does at
all. (Please refer to the Emacs Lisp manual for more information.)

Luckily we don't need a variable at all, but if we did, then we
would use `let' to introduce it.

Furthermore `system-type's value is a symbol not a string, wrapping
the condition in `not' needlessly complicates things, and the lines
are too long.

1: beda920037a36084d0d312cf2d1133314f68feba
   [Fix] Gross hack to work around problem on *BSD