Silence byte-compiler
In truth `looking-back' does not (yet) require the second argument,
but since 5161c9ca6a6107da30d411fb2ad72e01d08e5704 it claims that
it does.

Instead of merely silencing this warning it would be better to
provide a non-nil value for LIMIT.  The reason that this function
is now advertised as requiring two arguments is that it may perform
horribly if no proper limit is specified.