Add a fallback shell command
This commit introduces a new `defcustom` that configures a fallback
shell executable which is used if the selected shell
(e.g. `vterm-shell`) is not executable.

The fallback is especially useful if the `vterm` is opened on a remote
host over ssh/tramp which does not have the configured `vterm-shell`
installed, but the user does not want to configure a different shell for
all ssh hosts, as some may have the shell in `vterm-shell` installed.