Use same shell as TRAMP on remotes.
If emacs-vterm is started on a remote, use the same shell as TRAMP. That
will guarantee that the shell exists on the remote.

A shell could be chosen using `vterm-tramp-shells', but it tied to a
specific TRAMP method. However, the same method (ssh) could be used to
connected to different remotes, each with different default shell. So,
we could use TRAMP to get the shell on per-remote basis.