Don't allow moving cursor below the last prompt via mouse
Without this patch, I am able to click somewhere below the last
prompt line and move the cursor there. I don't think this was an
intended feature, it feels more like a weird quirk.

We don't have any function telling us the point or line number of the
last prompt, and it doesn't seem to be easy to implement, so I
needed to workaround this.