Set the vterm buffer's major mode before displaying the buffer.
This ensures that display-buffer actions that condition on a buffer's
major mode (see the documentation of `display-buffer-alist` and
`buffer-match-p`) properly catch the buffer when it is displayed for the
first time.

This aligns with the behavior of other terminal emulator
modes (e.g. term and ansi-term) and comint-derived modes (e.g. ielm,
SQLi, and inferior Python).