Fixes #729: Ignore libvterm string with length -1
libvterm sometimes gives us a string fragment with length 1073741823 [1]. When
this happens, we attempt to read 1073741823 bytes, overrunning the buffer and
causing a segfault.

Fix the crash by adding a check for -1. Because the number is not exactly -1
(because VTermStringFragment.len is a 30-bit bit field at least in my copy of
libvterm), do some C trickery to get the -1 value for comparison.

[1] This -1 length might be a bug in libvterm. I did not investigate further.